# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API VIDA",
        "description": "API para la aplicación VIDA",
        "contact": {
            "email": "asd@asd.asd"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Server API VIDA"
        }
    ],
    "paths": {
...
        "/Colegios": {
            "get": {
                "tags": [
                    "Colegios"
                ],
                "description": "Muestra todos los colegios",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "Todos los colegios"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Colegios"
                ],
                "description": "Crea un nuevo colegio en la base de datos",
                "operationId": "store",
                "requestBody": {
                    "description": "Datos a ingresar en la base de datos",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "idColegio": {
                                        "description": "ID del colegio a ingresar en la base de datos",
                                        "type": "integer",
                                        "format": "int64"
                                    },
                                    "idMunicipio": {
                                        "description": "ID del municipio en el que se encuentra el colegio",
                                        "type": "integer",
                                        "format": "int64"
                                    },
                                    "Localidad": {
                                        "description": "Nombre de la localidad en la que se encuentra el colegio",
                                        "type": "string"
                                    },
                                    "idProvincia": {
                                        "description": "ID de la provincia en la que se encuentra el colegio",
                                        "type": "integer",
                                        "format": "int64"
                                    },
                                    "Provincia": {
                                        "description": "Nombre de la provincia en la que se encuentra el colegio",
                                        "type": "string"
                                    },
                                    "Denominacion_generica": {
                                        "description": "Colegio de educacion infantil, Centro privado de Secundaria, etc",
                                        "type": "string"
                                    },
                                    "Denominacion_especifica": {
                                        "description": "Nombre del centro",
                                        "type": "string"
                                    },
                                    "Naturaleza": {
                                        "description": "Centro privado o público",
                                        "type": "string"
                                    },
                                    "Domicilio": {
                                        "description": "Direccion en la que se encuentra el colegio",
                                        "type": "string"
                                    },
                                    "C_Postal": {
                                        "description": "Código postal del colegio",
                                        "type": "integer",
                                        "format": "int64"
                                    },
                                    "Telefono": {
                                        "description": "Telefono del colegio",
                                        "type": "integer",
                                        "format": "int64"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Colegio agregado correctamente"
                    },
                    "500": {
                        "description": "Ha ingresado uno o varios campos erroneamente o hay un problema con el servidor"
                    }
                }
            }
        },
        "/Colegios/{idColegio}": {
            "get": {
                "tags": [
                    "Colegios"
                ],
                "description": "Muestra la informacion del colegio ",
                "operationId": "App\\Http\\Controllers\\ColegiosController::show",
                "parameters": [
                    {
                        "name": "idColegio",
                        "in": "path",
                        "description": "ID del colegio en el que buscar informacion",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Muestra informacion del colegio"
                    },
                    "404": {
                        "description": "No se ha encontrado el colegio"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Colegios"
                ],
                "description": "Actualiza el colegio elegido en la base de datos",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "idColegio",
                        "in": "path",
                        "description": "id del colegio a editar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Datos a modificar en la base de datos",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "idColegio": {
                                        "description": "ID del colegio a ingresar en la base de datos",
                                        "type": "integer",
                                        "format": "int64"
                                    },
                                    "idMunicipio": {
                                        "description": "ID del municipio en el que se encuentra el colegio",
                                        "type": "integer",
                                        "format": "int64"
                                    },
                                    "Localidad": {
                                        "description": "Nombre de la localidad en la que se encuentra el colegio",
                                        "type": "string"
                                    },
                                    "idProvincia": {
                                        "description": "ID de la provincia en la que se encuentra el colegio",
                                        "type": "integer",
                                        "format": "int64"
                                    },
                                    "Provincia": {
                                        "description": "Nombre de la provincia en la que se encuentra el colegio",
                                        "type": "string"
                                    },
                                    "Denominacion_generica": {
                                        "description": "Colegio de educacion infantil, Centro privado de Secundaria, etc",
                                        "type": "string"
                                    },
                                    "Denominacion_especifica": {
                                        "description": "Nombre del centro",
                                        "type": "string"
                                    },
                                    "Naturaleza": {
                                        "description": "Centro privado o público",
                                        "type": "string"
                                    },
                                    "Domicilio": {
                                        "description": "Direccion en la que se encuentra el colegio",
                                        "type": "string"
                                    },
                                    "C_Postal": {
                                        "description": "Código postal del colegio",
                                        "type": "integer",
                                        "format": "int64"
                                    },
                                    "Telefono": {
                                        "description": "Telefono del colegio",
                                        "type": "integer",
                                        "format": "int64"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Se ha modificado el colegio correctamente"
                    },
                    "500": {
                        "description": "Ha ingresado uno o varios campos erroneamente o hay un problema con el servidor"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Colegios"
                ],
                "description": "Elimina el colegio elegido",
                "operationId": "App\\Http\\Controllers\\ColegiosController::destroy",
                "parameters": [
                    {
                        "name": "idColegio",
                        "in": "path",
                        "description": "id del colegio a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Se ha eliminado correctamente"
                    },
                    "404": {
                        "description": "No se ha encontrado el colegio con el ID seleccionado"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
...
        "/Municipios/{idMunicipio}": {
            "get": {
                "tags": [
                    "Municipios"
                ],
                "description": "Muestra la informacion del municipio",
                "operationId": "App\\Http\\Controllers\\MunicipiosController::show",
                "parameters": [
                    {
                        "name": "idMunicipio",
                        "in": "path",
                        "description": "id del municipio en el que buscar informacion",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Muestra la informacion del municipio"
                    },
                    "404": {
                        "description": "No se ha encontrado el municipio"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Municipios"
                ],
                "description": "modifica un municipio en la base de datos",
                "operationId": "store",
                "parameters": [
                    {
                        "name": "idMunicipio",
                        "in": "path",
                        "description": "id del municipio a modificar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Datos a ingresar en la base de datos",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "CODMU": {
                                        "description": "ID del municipio a ingresar en la base de datos",
                                        "type": "integer",
                                        "format": "int64"
                                    },
                                    "NOMBRE": {
                                        "description": "Nombre del municipio",
                                        "type": "string"
                                    },
                                    "CODPROV": {
                                        "description": "ID de la provincia a la que pertenece",
                                        "type": "integer",
                                        "format": "int64"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Municipio modificado correctamente"
                    },
                    "500": {
                        "description": "Ha ingresado uno o varios campos erroneamente o hay un problema con el servidor"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Municipios"
                ],
                "description": "Elimina el municipio elegido",
                "operationId": "App\\Http\\Controllers\\MunicipiosController::destroy",
                "parameters": [
                    {
                        "name": "idMunicipio",
                        "in": "path",
                        "description": "id del municipio a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Se ha eliminado correctamente"
                    },
                    "404": {
                        "description": "No se ha encontrado el colegio con el ID seleccionado"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
...
        "/Provincias/{idProvincia}": {
            "get": {
                "tags": [
                    "Provincias"
                ],
                "description": "Muestra la informacion de la provincia",
                "operationId": "App\\Http\\Controllers\\ProvinciasController::show",
                "parameters": [
                    {
                        "name": "idProvincia",
                        "in": "path",
                        "description": "id de la provincia en la que buscar informacion",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Muestra la informacion de la provincia"
                    },
                    "404": {
                        "description": "No se ha encontrado la provincia"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Provincias"
                ],
                "description": "Crea una nueva provincia en la base de datos",
                "operationId": "store",
                "parameters": [
                    {
                        "name": "idProvincia",
                        "in": "path",
                        "description": "id de la provincia a modificar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Datos a modificar en la base de datos",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "CODPROV": {
                                        "description": "ID de la provincia a ingresar en la base de datos",
                                        "type": "integer",
                                        "format": "int64"
                                    },
                                    "NOMBRE": {
                                        "description": "Nombre de la provincia",
                                        "type": "integer",
                                        "format": "int64"
                                    },
                                    "CODAUTO": {
                                        "description": "Codigo de la comunidad autonoma a la que pertenece",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Provincia agregada correctamente"
                    },
                    "500": {
                        "description": "Ha ingresado uno o varios campos erroneamente o hay un problema con el servidor"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Provincias"
                ],
                "description": "Elimina la provincia elegida",
                "operationId": "App\\Http\\Controllers\\ProvinciasController::destroy",
                "parameters": [
                    {
                        "name": "idProvincia",
                        "in": "path",
                        "description": "id de la provincia a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Se ha eliminado correctamente"
                    },
                    "404": {
                        "description": "No se ha encontrado el colegio con el ID seleccionado"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
...
    },
...
    "security": [
        []
    ]
}